fix(webapp): export versionfooterprops to fix storybook type portability
Export the VersionFooterProps interface to resolve TS4023 'cannot be named'
error that occurs when TypeScript generates declaration files for stories
using 'satisfies Meta<typeof Component>' pattern.

This is a known Storybook 10 type portability issue that requires component
props interfaces to be exported when used with the satisfies operator in
stories.